Avoid polluting /tmp when running tests

+++ test/utilstest.py
@@ -348,21 +348,24 @@ def Rwp(obt, ref, comment="Rwp"):
     return numpy.sqrt(((big_delta[non_null]) ** 2 / ((big_mean[non_null]) ** 2)).sum())
 
 
-def recursive_delete(strDirname):
+def recursive_delete(dirname):
     """
     Delete everything reachable from the directory named in "top",
     assuming there are no symbolic links.
     CAUTION:  This is dangerous!  For example, if top == '/', it
     could delete all your disk files.
-    @param strDirname: top directory to delete
-    @type strDirname: string
+
+    @param dirname: top directory to delete
+    @type dirname: string
     """
-    for root, dirs, files in os.walk(strDirname, topdown=False):
+    if not os.path.isdir(dirname):
+        return
+    for root, dirs, files in os.walk(dirname, topdown=False):
         for name in files:
             os.remove(os.path.join(root, name))
         for name in dirs:
             os.rmdir(os.path.join(root, name))
-    os.rmdir(strDirname)
+    os.rmdir(dirname)
 
 getLogger = UtilsTest.get_logger
